code-move
* move `ReplicaState`, `ReplicaTaskConfig` , and `replica_task` into
  replica.rs
* move ` Instance`, `ActiveInstance`, `PendingPeek`, `ReportedUppers`,
  and `FrontierBounds` into instance.rs
